fix: log as JSON
At the moment we log as plain text, which has a number of disadvantages:

* Limited structure
* Exceptions/multiline logs are treated as separate lines by e.g.
  cloudwatch, ELK, making it harder to debug.
* Hard to add extra data in a consistent way.

By switching to JSON, we can ensure that each line in our logs
encapsulates a specific action/record entirely, and it's very easy to
add extra fields. Our ELK stack will also be easier to use/filter/read
as it supports json log format natively.