fix: search box placeholder
Refactor the `search-v2` JS file so that the features it provides need
to be initialised in the template using the file. This lets us reuse the
focus toggle for the search box placeholder in the dataset finder index
page.

Also fixes a bug that would sometimes cause the search page to get
locked in a greyed out state if the stored active element has been
cleared up.